画像モーダルの状態・ロジック管理と初期画像設定をカスタムフックに切り出し
