+++ frontend/src/App.js
@@ -6,8 +6,7 @@ import theme from "./components/styles/theme";
 import GlobalStyle from "./components/GlobalStyle"
 
 import { BrowserRouter, Routes, Route, useLocation } from "react-router-dom";
-//import { Home, CryptidList, Cryptid } from "./pages";
-import Cryptid from "./pages/Cryptid";
+import Cryptid from "./pages/Cryptid/Cryptid";
 import {HomeContainer, CryptidsContainer} from './containers';
 
 import { Header, Footer } from "./components/layouts";
