--- frontend/src/pages/xxxxxxxxxxxindex.js
@@ -1,3 +1,0 @@
-export { default as Page } from "./Home/Page";
-export { default as CryptidList } from "./CryptidList";
-export { default as Cryptid } from "./Cryptid";
